`Unimplemented` should wrap a base clock
Currently, "unimplemented" clocks are simply immediate clocks that emit
XCTest failures, but the immediacy can be surprising. In general
"unimplemented" versions of dependencies (UUID, Date, etc.) simply call
out to the live dependency, but fail, so we should provide a way to do
the same. Now we can, for example:

```swift
UnimplementedClock(.continuous)
```